+++ src/Products/MeetingCommunes/adapters.py
@@ -700,7 +700,7 @@ class CustomMeetingConfig(MeetingConfig):
         infos.update(extra_infos)
 
         # disable FINANCE_ADVICES_COLLECTION_ID excepted for 'meeting-config-college' and 'meeting-config-bp'
-        if cfg.getId() not in ('meeting-config-college', 'meeting-config-bp'):
+        if cfg.getId() not in ('meeting-config-college', 'meeting-config-zcollege', 'meeting-config-bp'):
             infos[FINANCE_ADVICES_COLLECTION_ID]['active'] = False
 
         # add some specific searches while using 'meetingadvicefinances'
@@ -910,6 +910,8 @@ class MeetingCommunesWorkflowActions(MeetingWorkflowActions):
            MeetingConfig.initItemDecisionIfEmptyOnDecide is True, we
            initialize the decision field with content of Title+Description
            if decision field is empty.'''
+        # call parent's doDecide
+        super(MeetingCommunesWorkflowActions, self).doDecide(stateChange)
         if self.cfg.getInitItemDecisionIfEmptyOnDecide():
             for item in self.context.get_items():
                 # If deliberation (motivation+decision) is empty,